+++ src/venmo/internal/ui-elements/error-view.js
@@ -48,7 +48,7 @@ var ErrorView = /** @class */ (function (_super) {
         return container;
     };
     ErrorView.prototype.getStyleConfig = function () {
-        return "\n      #venmo-error-view {\n        display: none;\n        flex-direction: column;\n        align-items: center;\n        justify-content: center;\n        text-align: center;\n        color: #FFFFFF;\n      }\n\n      #venmo-error-view.active {\n        display: flex;\n      }\n\n      #venmo-error__button {\n        cursor: pointer;\n        margin-top: 30px;\n        background: linear-gradient(0deg, rgba(0, 0, 0, 0.1), rgba(0, 0, 0, 0.1)), #0074DE;\n        border-radius: 24px;\n        padding: 14px;\n        min-width: 300px;\n        font-size: 18px;\n        font-weight: bold;\n        line-height: 24px;\n        color: #FFFFFF;\n      }\n    ";
+        return "\n      #venmo-error-view {\n        display: none;\n        flex-direction: column;\n        align-items: center;\n        justify-content: center;\n        text-align: center;\n        color: #FFFFFF;\n        background: #2F3033;\n        box-shadow: 0px 0px 8px rgba(0, 0, 0, 0.4);\n        border-radius: 8px;\n        padding: 0 10px;\n      }\n\n      #venmo-error-view.active {\n        display: flex;\n      }\n\n      #venmo-error__button {\n        cursor: pointer;\n        margin-top: 30px;\n        background: linear-gradient(0deg, rgba(0, 0, 0, 0.1), rgba(0, 0, 0, 0.1)), #0074DE;\n        border-radius: 24px;\n        padding: 14px;\n        min-width: 300px;\n        font-size: 18px;\n        font-weight: bold;\n        line-height: 24px;\n        color: #FFFFFF;\n      }\n    ";
     };
     return ErrorView;
 }(base_1.default));