+++ src/venmo/internal/ui-elements/modal.js
@@ -40,9 +40,6 @@ var Modal = /** @class */ (function (_super) {
         _this.closeIconContainerElement = _this.$("#close-icon-container");
         close_icon_1.default.create({
             container: _this.$("#close-icon-container"),
-            onClick: function () {
-                _this.close();
-            },
         });
         return _this;
     }
@@ -56,11 +53,11 @@ var Modal = /** @class */ (function (_super) {
         this.viewBoxesElement.classList.toggle("is-flipped");
     };
     Modal.prototype.showFrontFace = function () {
-        this.frontView.reset();
+        this.resetViews();
         this.viewBoxesElement.classList.remove("is-flipped");
     };
     Modal.prototype.showBackFace = function () {
-        this.backView.reset();
+        this.resetViews();
         this.viewBoxesElement.classList.add("is-flipped");
     };
     Modal.prototype.displayQRCode = function (url) {
@@ -87,20 +84,33 @@ var Modal = /** @class */ (function (_super) {
     };
     Modal.prototype.reset = function () {
         this.authorized = false;
+        this.hide();
         this.showFrontFace();
-        this.backView.reset();
         this.closeIconContainerElement.classList.remove("hidden");
+    };
+    Modal.prototype.show = function () {
+        this.$("#outer-container").classList.remove("hidden");
+    };
+    Modal.prototype.hide = function () {
+        this.$("#outer-container").classList.add("hidden");
+    };
+    Modal.prototype.resetViews = function () {
+        this.frontView.reset();
+        this.backView.reset();
     };
     Modal.prototype.constructElement = function () {
         var _this = this;
         var modal = document.createElement("div");
         modal.id = "venmo-desktop-modal";
-        modal.innerHTML = "\n    <div id=\"outer-container\">\n      <div id=\"close-icon-container\"></div>\n      <div id=\"view-boxes\">\n        <div id=\"view-boxes-container\">\n        </div>\n      </div>\n    </div>\n    ";
+        modal.innerHTML = "\n    <div id=\"close-icon-container\"></div>\n    <div id=\"outer-container\">\n      <div id=\"view-boxes\">\n        <div id=\"view-boxes-container\">\n        </div>\n      </div>\n    </div>\n    ";
         modal_backdrop_1.default.create({
             container: document.body,
             onClick: function () {
                 _this.close();
             },
+        });
+        modal.addEventListener("click", function () {
+            _this.close();
         });
         window.addEventListener("keydown", function (event) {
             if (event.key === "Escape") {
@@ -110,7 +120,7 @@ var Modal = /** @class */ (function (_super) {
         return modal;
     };
     Modal.prototype.getStyleConfig = function () {
-        return "\n      #venmo-desktop-modal {\n        display: flex;\n        align-items: center;\n        justify-content: center;\n        height: 100vh;\n      }\n\n      #close-icon-container.hidden {\n        display: none;\n      }\n\n      #outer-container {\n        position: relative;\n        top: 0;\n        width: 720px;\n        max-width: 95%;\n        height: 480px;\n        background: #2F3033;\n        box-shadow: 0px 0px 8px rgba(0, 0, 0, 0.4);\n        border-radius: 16px;\n        perspective: 840px;\n        animation: 1s drop;\n      }\n\n      #view-boxes {\n        display: flex;\n        align-items: center;\n        justify-content: center;\n        height: 100%;\n        width: 100%;\n        perspective: 1000;\n        transition: transform 1s;\n        transform-style: preserve-3d;\n      }\n\n      #view-boxes-container {\n        display: flex;\n        align-items: center;\n        justify-content: center;\n        height: 100%;\n        width: 100%;\n        font-family: sans-serif;\n        font-style: normal;\n        font-weight: 100;\n        perspective: 1000;\n      }\n\n      .view-box {\n        display: flex;\n        max-width: 95%;\n        width: 280px;\n        height: 321px;\n        -webkit-backface-visibility: hidden; /* Safari */\n        backface-visibility: hidden;\n        transition: transform 1s;\n        transform-style: preserve-3d;\n      }\n    ";
+        return "\n      #venmo-desktop-modal {\n        display: flex;\n        align-items: center;\n        justify-content: center;\n        height: 100vh;\n      }\n\n      #close-icon-container.hidden {\n        display: none;\n      }\n\n      #outer-container {\n        position: absolute;\n        top: 0;\n        bottom: 0;\n        width: 100%;\n        perspective: 840px;\n        animation: 1s drop;\n      }\n\n      #outer-container.hidden {\n        display: none;\n      }\n\n      #view-boxes {\n        display: flex;\n        align-items: center;\n        justify-content: center;\n        height: 100%;\n        width: 100%;\n        perspective: 1000;\n        transition: transform 1s;\n        transform-style: preserve-3d;\n      }\n\n      #view-boxes-container {\n        display: flex;\n        align-items: center;\n        justify-content: center;\n        height: 100%;\n        width: 100%;\n        font-family: sans-serif;\n        font-style: normal;\n        font-weight: 100;\n        perspective: 1000;\n      }\n\n      .view-box {\n        display: flex;\n        max-width: 95%;\n        width: 280px;\n        height: 321px;\n        -webkit-backface-visibility: hidden; /* Safari */\n        backface-visibility: hidden;\n        transition: transform 1s;\n        transform-style: preserve-3d;\n      }\n    ";
     };
     return Modal;
 }(base_1.default));